kanban-board-backend/pyproject.toml
[project]
name = "kanban-board-backend"
version = "0.1.0"
description = "Implementação de testes de um kanban board"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "playwright>=1.58.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
]
